@@ -28,16 +28,13 @@ class ParticlesGame extends BaseGame {
   /// Defines the lifespan of all the particles in these examples
   final sceneDuration = const Duration(seconds: 1);
 
-  Vector2 cellSize;
-  Vector2 halfCellSize;
+  Vector2 get cellSize => size / gridSize;
+  Vector2 get halfCellSize => cellSize / 2;
 
   @override
   Future<void> onLoad() async {
     await images.load('zap.png');
     await images.load('boom.png');
-
-    cellSize = size / gridSize;
-    halfCellSize = cellSize * .5;
 
     // Spawn new particles every second
     Timer.periodic(sceneDuration, (_) => spawnParticles());
@@ -78,8 +75,7 @@ class ParticlesGame extends BaseGame {
       final particle = particles.removeLast();
       final col = particles.length % gridSize;
       final row = (particles.length ~/ gridSize).toDouble();
-      final cellCenter =
-          (cellSize.clone()..multiply(Vector2(col, row))) + (cellSize * .5);
+      final cellCenter = (cellSize..multiply(Vector2(col, row))) + halfCellSize;
 
       add(
         // Bind all the particles to a [Component] update
@@ -218,7 +214,7 @@ class ParticlesGame extends BaseGame {
             Colors.red,
             Colors.blue,
             particle.progress,
-          ),
+          )!,
       ),
     );
   }
@@ -242,14 +238,14 @@ class ParticlesGame extends BaseGame {
               Colors.red,
               Colors.blue,
               steppedProgress,
-            ),
+            )!,
         );
       },
     );
   }
 
   /// Particle which is used in example below
-  Particle reusablePatricle;
+  Particle? reusablePatricle;
 
   /// A burst of white circles which actually using a single circle
   /// as a form of optimization. Look for reusing parts of particle effects
@@ -261,7 +257,7 @@ class ParticlesGame extends BaseGame {
       generator: (i) => MovingParticle(
         curve: Interval(rnd.nextDouble() * .1, rnd.nextDouble() * .8 + .1),
         to: randomCellOffset() * .5,
-        child: reusablePatricle,
+        child: reusablePatricle!,
       ),
     );
   }
@@ -277,7 +273,7 @@ class ParticlesGame extends BaseGame {
   }
 
   /// Particle which is used in example below
-  Particle reusableImageParticle;
+  Particle? reusableImageParticle;
 
   /// A single [imageParticle] is drawn 9 times
   /// in a grid within grid cell. Looks as 9 particles
@@ -298,7 +294,7 @@ class ParticlesGame extends BaseGame {
           (i % perLine) * colWidth - halfCellSize.x + imageSize,
           (i ~/ perLine) * rowHeight - halfCellSize.y + imageSize,
         ),
-        child: reusableImageParticle,
+        child: reusableImageParticle!,
       ),
     );
   }
