@@ -6,32 +6,32 @@ import 'package:flutter/material.dart';
 
 /// Includes an example including advanced detectors
 class MultitapGame extends BaseGame with MultiTouchTapDetector {
-  static final _whitePaint = BasicPalette.white.paint;
-  static final _size = Vector2.all(50);
+  static final whitePaint = BasicPalette.white.paint;
+  static final tapSize = Vector2.all(50);
 
-  final Map<int, Rect> _taps = {};
+  final Map<int, Rect> taps = {};
 
   @override
   void onTapDown(int pointerId, TapDownDetails details) {
-    _taps[pointerId] =
-        details.globalPosition.toVector2().toPositionedRect(_size);
+    taps[pointerId] =
+        details.globalPosition.toVector2().toPositionedRect(tapSize);
   }
 
   @override
   void onTapUp(int pointerId, _) {
-    _taps.remove(pointerId);
+    taps.remove(pointerId);
   }
 
   @override
   void onTapCancel(int pointerId) {
-    _taps.remove(pointerId);
+    taps.remove(pointerId);
   }
 
   @override
   void render(Canvas canvas) {
     super.render(canvas);
-    _taps.values.forEach((rect) {
-      canvas.drawRect(rect, _whitePaint);
+    taps.values.forEach((rect) {
+      canvas.drawRect(rect, whitePaint);
     });
   }
 }
