@@ -5,8 +5,8 @@ import 'package:flame/game.dart';
 import 'package:flame/palette.dart';
 import 'package:flutter/material.dart';
 
-final regular = TextConfig(color: BasicPalette.white.color);
-final tiny = regular.withFontSize(12.0);
+final _regular = TextConfig(color: BasicPalette.white.color);
+final _tiny = _regular.withFontSize(12.0);
 
 final _white = Paint()
   ..color = BasicPalette.white.color
@@ -16,7 +16,7 @@ class MyTextBox extends TextBoxComponent {
   MyTextBox(String text)
       : super(
           text,
-          config: tiny,
+          config: _tiny,
           boxConfig: TextBoxConfig(
             timePerChar: 0.05,
             growingBox: true,
@@ -43,20 +43,20 @@ class TextGame extends BaseGame {
   @override
   Future<void> onLoad() async {
     add(
-      TextComponent('Hello, Flame', config: regular)
+      TextComponent('Hello, Flame', config: _regular)
         ..anchor = Anchor.topCenter
         ..x = size.x / 2
         ..y = 32.0,
     );
 
     add(
-      TextComponent('center', config: tiny)
+      TextComponent('center', config: _tiny)
         ..anchor = Anchor.center
         ..position = size / 2,
     );
 
     add(
-      TextComponent('bottomRight', config: tiny)
+      TextComponent('bottomRight', config: _tiny)
         ..anchor = Anchor.bottomRight
         ..position = size,
     );
