@@ -9,7 +9,8 @@ class MySpriteBatchComponent extends SpriteBatchComponent
     with HasGameRef<SpritebatchAutoLoadGame> {
   @override
   Future<void> onLoad() async {
-    spriteBatch = await gameRef.loadSpriteBatch('boom.png');
+    final spriteBatch = await gameRef!.loadSpriteBatch('boom.png');
+    this.spriteBatch = spriteBatch;
 
     spriteBatch.add(
       source: const Rect.fromLTWH(128 * 4.0, 128 * 4.0, 64, 128),
@@ -25,14 +26,14 @@ class MySpriteBatchComponent extends SpriteBatchComponent
       color: Colors.redAccent,
     );
 
+    final size = gameRef!.size;
     const num = 100;
     final r = Random();
     for (var i = 0; i < num; ++i) {
       final sx = r.nextInt(8) * 128.0;
       final sy = r.nextInt(8) * 128.0;
-      final x = r.nextInt(gameRef.size.x.toInt()).toDouble();
-      final y =
-          r.nextInt(gameRef.size.y ~/ 2).toDouble() + gameRef.size.y / 2.0;
+      final x = r.nextInt(size.x.toInt()).toDouble();
+      final y = r.nextInt(size.y ~/ 2).toDouble() + size.y / 2.0;
       spriteBatch.add(
         source: Rect.fromLTWH(sx, sy, 128, 128),
         offset: Vector2(x - 64, y - 64),
